bingo-generator/requirements.txt
reportlab>=4.0.0
qrcode[pil]>=7.4.0
pillow>=10.0.0
